README.md
People are rarely aware of the environmental and social implications of the products we buy. Buy Better strives to make conscious consumerism the status quo. Navigate through a wide range of products that are pre-vetted as fair-trade, eco-friendly, ethical or organic. Buy Better is a mobile app that provides a full fledged online shopping experience you can feel good about. 

Buy the things you love and while reducing your footprint and making a having a positive impact on the lives of the people behind the product. It’s that easy. Start here. Buy Better.